gs-appstream: Fix handling of language kudo for non-trivial locales
We can use the g_get_locale_variants() function from GLib to produce all
the combinations of language, territory, codeset and modifier from the
currently-set locale. This simplifies the code a little. Previously, the
code would fail to correctly get the language if the user’s locale had a
language and modifier, but no territory. (I don’t have any locales like
that on my system, but it’s possible.)

Signed-off-by: Philip Withnall <withnall@endlessm.com>