Fix warning when refining an app with <content_rating> without type attribute
It may happen that some apps have <content_rating> defined but without the
"type" attribute set (e.g. "com.sigmyne.crush").

This causes a warning when trying to compare the type with the known
supported values.

This patch fixes this issue by ignoring the content rating when type is
not set. It also avoids some unnecessary processing by changing the order
of the conditionals, by first checking whether the type is set/valid
before doing any other processing based on it.