Fix issues with GsAppListFilterFlags and GObject properties on 32 bits systems
GsPluginJob "dedupe-flags" property expects GsAppListFilterFlags as uint64.

This causes a crash on 32 bits systems when passing GsAppListFilterFlags to
g_object_new(), as the type is not properly interpreted as uint64.

Lets force the type to be uint64 so as to avoid a need for casting all
places using g_object_new and GsAppListFilterFlags.

Reworked a little by Philip Withnall.