gs-app: Use g_object_notify_by_pspec
`g_object_notify()` actually takes a global lock to look up the
`GParamSpec` that corresponds to the given property name. It's not a
huge performance hit, but it's easily avoidable by using the
`_by_pspec()` variant.

Furthermore, we can prevent a string copy on each
`gs_app_queue_notify()` because we don't need to `strdup` the property
name. We even don't need to to ref the `GParamSpec`, since we know that
at least one reference to a `GsApp` will be kept (due to the
`g_object_ref`).