gs-details-page: Rework age ratings handling to avoid hard-coded list
The list of OARS categories in the code to build age ratings was not in
sync with the latest OARS release (v1.1). In particular, it was missing
`drugs-tobacco` from v1.0, and all of the categories from v1.1.

Fix that, and try to avoid it happening again in future, by querying
`gs-content-rating.c` for the list of categories, and only maintaining lists
of categories we want to group together in GNOME Software.

Signed-off-by: Philip Withnall <withnall@endlessm.com>