flatpak: Ensure use of the installed_refs cache is thread-safe
Just like the `broken_remotes` cache, the `installed_refs` cache is
shared between multiple threads. Previously, there was no mutex
protecting access to it, so one thread could replace it while another
was using it. Or, more typically (which is how I spotted this), two
threads would race to call `flatpak_installation_list_installed_refs()`
to populate it; one would win, and the other thread’s value would be
leaked.

This was resulting in 80KB of leaks after a few minutes of using
gnome-software.

Signed-off-by: Philip Withnall <withnall@endlessm.com>