flatpak: Only ignore estimated progress if >10%
Flatpak may set the "estimating" flag on a progress report if the value
is bogus, or it may set it if it is downloading metadata which it
arbitrarily decides comprises the first 5% of an install operation.
Ignoring estimated progress therefore means we get no progress changes
for the first 10% of an app install (installing the app and its locale
extension). Instead, only ignore the estimated progress if it is an
obviously bogus value of >10%.