core: Use GS_APP_QUIRK_HIDE_EVERYWHERE instead of Blacklisted category
Hide things based on the new quirk rather than the category. Put some
conversions in place in the `gs-appstream` plugin to apply the quirk if
the category is present (for backwards compatibility). Add a new
metadata key to apply the quirk.

Applications will now be blacklisted if they contain this in their
appdata:
```
<metadata>
  <value key="GnomeSoftware::quirks::hide-everywhere">true</value>
</metadata>
```

And they will also be blacklisted if they contain this, which is now
deprecated:
```
<categories>
  <category>Blacklisted</category>
</categories>
```

Signed-off-by: Philip Withnall <withnall@endlessm.com>

Fixes: #696