gs-details-page: Reset star rating when loading an unrated app
Previously, if the details page was loaded for an app with some ratings
(so an app rating of 0/5 to 5/5), and then loaded for an app with no
ratings (an unset app rating of -1/5), the ratings widget would continue
to show the app rating from the first app.

Fix that by unconditionally updating the widget, and marking it as
insensitive if the current app’s rating is unset.

Signed-off-by: Philip Withnall <withnall@endlessm.com>