gs-review-dialog: Fix suggested-action highlighting of Post button
The ‘Post’ button on the review dialog is meant to be highlighted in
blue as a suggested action, but that wasn’t happening due to logic in
`GtkDialog`’s `update_suggested_action()` removing that style class for
any header bar widget which was not the default action.

Fix that by making the ‘Post’ button the default action. Ensure it’s
insensitive by default.

Signed-off-by: Philip Withnall <withnall@endlessm.com>