Mark featured AppStream snippets as append-merge components
The components in collection XML have to be marked with a merge mode
explicitly if they are expected to be joined together with other
component metadata. Otherwise they will be read as separate new, but
invalid, generic components.
This change ensures the GNOME Software metadata additions get registered
no matter which AppStream implementation is used.