snap: Use same auth data as the command line client
Previously we have a graphical means of logging into the snap store, but this
was removed in 145bb82. This change means we now use the same auth data as
the command line snap client where previously GNOME Software used it's own
auth data store. For a user this requires them to have run 'snap login' from the
command line at some point.