source selector: Fix vertical alignment of widgets.
Stop widgets from expanding vertically and instead set the vertical alignment of
the grid which resolves the reason they were set to expand.

The existing code was observed causing the header bar to vertically expand,
which may have been an issue in that particular version of GTK. The fix is still
applicable as these widgets should never expand vertically.