upgrade banner: Show the correct button when NEEDS_REBOOT quirk is set
The logic here is that all distro upgrades need a reboot to *use* it,
but only some need a reboot to *deploy* it. We use the
GS_APP_QUIRK_NEEDS_REBOOT to denote distro upgrades that need to reboot
to deploy it, e.g. PackageKit distro upgrades or rpm-ostree rebases.