malcontent: Reload the app filter when it changes
All the code to reload all the apps and re-filter them was in place and
working, but I forgot to actually reload the filter when it changed,
leading to re-filtering the apps using the old filter.

This fixes reloading apps when the user’s app filter is changed in the
separate parental controls app (`malcontent-control`).

Signed-off-by: Philip Withnall <withnall@endlessm.com>