flatpak: Properly check if an app is a runtime
In gs_flatpak_set_kind_from_flatpak(), we use AS_APP_KIND_LOCALIZATION
and AS_APP_KIND_GENERIC in addition to AS_APP_KIND_RUNTIME, all for
things which from Flatpak's perspective are runtimes (their ref starts
with "runtime/"). So change checks in a few places to correctly identify
Flatpak runtimes instead of checking for AS_APP_KIND_RUNTIME, which has
the effect of ensuring that applications properly show as needing an
update when their locale extension needs an update, e.g. because the
user chose a language other than the one the locale extensions already
have.

Together with an update of libflatpak to 1.7.3 when that is released,
this commit addresses the issue described in this MR:
https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/430

Closes: #539