Don't offer Fedora distro upgrades if we don't have correct artwork
Make sure we have the artwork for the distro upgrade banner before
offering an upgrade. In practice, this means that we don't offer a
distro upgrade at all unless the user has first updated gnome-software
to the version available from in the updates repo.

This comes with a slight tradeoff that some people who never install
regular updates will not get the distro upgrade notification at all
after this change.