flatpak: don't show as legacy display without x11
Currently, every Flathub app is show as using a "legacy display manager"
whether they use x11 with --socket=x11, or wayland with
--socket=wayland and --socket=fallback-x11.

However, the fallback-x11 enables X11 only if the host doesn't support
anything else than x11, so strictly speaking, it's the host that uses
x11 and not usually the app.

So this commit stops from showing this warning if using x11 as fallback.