tests: Avoid using a hard-coded directory for temporary test data
This means that the tests cannot be run in parallel, since they will
race over the common cache files in `/var/tmp/self-test`. This is the
cause of a number of CI test failures, as Meson runs tests in parallel
by default (which is the correct thing to do — tests should be
self-contained).

Instead, create a unique temporary directory for each test. Ideally,
this would be done per unit test (each `g_test_add()`), but that’s not
possible without rearchitecting the plugin loader so that it can be
reinitialised for each unit test. That can be done in future work. As a
workaround, clear the contents of the unique temporary directory at the
start of each unit test.

Signed-off-by: Philip Withnall <withnall@endlessm.com>