tests: Enable G_TEST_OPTION_ISOLATE_DIRS for all tests
This gives stronger guarantees that the tests are not using the user’s
real home directory, or system paths outside the build root, without the
test harness explicitly setting that up. This makes the tests less
likely to cause harm, and more portable and reproducable.

A few tests need paths for access to the system icon theme and the
shared-mime-info database to be set up, as they need those. The
shared-mime-info database is needed for detecting content types like
`.tar.gz` for decompression of appdata or firmware archives, for
example.

The new APIs in use come from GLib 2.60, but since the CI is still
running on Debian Stable (GLib 2.58), make the dependency optional for
now. That’s OK, as this is a ‘safety net’ rather than a hard requirement
of the tests at the moment.

Signed-off-by: Philip Withnall <withnall@endlessm.com>