odrs: Reduce size of critical sections
Previously, the ODRS mutex was being held while a 780KB JSON file
(`ratings.json`) was being parsed and processed. In another place in the
ODRS plugin, the mutex was being acquired and released in a loop.
Neither are conducive to good performance.

Reduce the size/frequency of the critical sections by moving the mutex
locking out of a loop, and atomically replacing the entire ratings array
when refreshing it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>