odrs: Use a single GArray for loaded ratings data
This saves more than 4000 allocations and around 1MB of heap memory.

Previously, the ratings data was inserted into a hash table, allocating
the key, value, and various internal allocations inside the hash table
in GLib.

With this change, ratings data is stored in one contiguous `GArray`
allocation, which reduces fragmentation and means that only one strdup
is needed per additional app in the ratings data.

Conversely, the new approach means that lookups of ratings data (when
refining apps) use a binary search of the array, rather than a hash
table lookup. This will be a little slower, but should not be noticeably
slower. (It’s hard to profile the exact effect, though.)

This relies on `g_array_binary_search()`, which was released in GLib
2.62. A copy is included here so gnome-software’s dependency doesn’t
have to be bumped.

Signed-off-by: Philip Withnall <withnall@endlessm.com>