Improve the heuristic for detecting old-style AppStream override files
The heuristic was trying to detect old-style AppStream override files
that have the following structure:

      <component type=desktop>
        <id>org.kde.amarok.desktop</id>
        <categories>
          <category>AudioVideo</category>
          <category>Featured</category>
        </categories>
      </component>

When it found one, it gave it the wildcard quirk to avoid leaking the
result to the installed apps list.

This however incorrectly tripped on old appdata files that didn't
specify name and relied on filling the missing name/summary/icon from
the desktop file.

Fix this by tightening the heuristic and also look for
<metadata_license> that none of the override files should have.

This fixes RHEL firefox package to correctly show up when clicking on
'Show Details' in GNOME Shell.