tree-wide: Replace usages of whitelist/blacklist
This commit replaces usages of the terms whitelist and blacklist with
allowlist and blocklist, which are both more inclusive and more precise.
This terminology change has been implemented across many projects.[1][2]

There is one place we can't remove the term blacklist because it is
there for backwards compatibility, in plugins/core/gs-appstream.c

This commit causes three types of API breaks and therefore should not be
backported to stable branches.
1) The plugin-whitelist and plugin-blacklist CLI parameters have been
   renamed. These do not show up in Debian Code Search anywhere.
2) The environment variables GNOME_SOFTWARE_PLUGINS_WHITELIST and
   GNOME_SOFTWARE_PLUGINS_BLACKLIST have been renamed. These are for
   debugging and also don't show up in Debian Code Search.
3) The hardcoded-blacklist plugin has been renamed which means the
   associated file is now libgs_plugin_hardcoded-blocklist.so which may
   require distributions to change their packaging.

[1] https://go-review.googlesource.com/c/go/+/236857/
[2] https://9to5google.com/2020/06/07/google-chrome-blacklist-blocklist-more-inclusive/