gs-app-row: Coalesce refresh operations where possible
Instead of rebuilding the entire widget state every time something
changes, schedule a refresh to happen in an idle callback. This allows
refreshes to be coalesced, which reduces the number of refreshes
happening overall. Previously, for example, `gs_app_row_refresh()` was
being called 3 times as part of constructing and setting up a new
`GsAppRow` in `GsInstalledPage`.

This speeds up populating the `GsInstalledPage`.

`gs_app_row_refresh()` has been removed as a public symbol for
`GsAppRow` since refreshes should be managed internally to the widget.
It was unused elsewhere.

This, in conjunction with the previous commit, decreases the time for
`gs_installed_page_get_installed_cb()` from 500ms to 300ms (for 114
apps), and spreads out the cost of setting up the widgets for each app
row over one or more subsequent main context iterations, which reduces
freezing of the UI.

Signed-off-by: Philip Withnall <withnall@endlessm.com>