fedora-pkgdb-collections: Memory map JSON files for parsing
Where possible, memory map any JSON files which are going to be parsed,
rather than loading the file into a new heap allocation, parsing it and
then freeing it. This saves a few large (temporary) allocations.

It uses a new API in the as-yet-unreleased json-glib 1.6, but doesn’t
have a hard dependency on it.

See https://gitlab.gnome.org/GNOME/json-glib/-/merge_requests/27

Signed-off-by: Philip Withnall <withnall@endlessm.com>