gs-plugin-loader: Add sysprof support for profiling jobs
This can be used with a command like:
```
sysprof-cli --gtk --use-trace-fd -- gnome-software
```
or from within GNOME Builder, to profile gnome-software. The changes
in this commit use the `sysprof-capture-3` library to write additional
timing data to the capture, showing the details of every plugin job and
vfunc call which happen in gnome-software.

This should make it easier to visualise what plugin jobs are happening,
and which ones are taking a long time.

This commit adds an optional dependency on `sysprof-capture-3`.

Signed-off-by: Philip Withnall <withnall@endlessm.com>