fwupd: Send our implemented feature set
At the moment we just blindly assume the capabilities of the front-end client
when installing firmware. We can somewhat work around this limitation by
requiring a new enough fwupd daemon version, but the GUI client software may be
much older than the fwupd version or just incomplete.

Clients that do not register features are assumed to be dumb and won't be
offered firmware that has a hard requirement on showing text or screenshots.