Do not use GsApp screenshot for transient screenshots
We currently support showing the user a screenshot which provides a more
friendly way to put a device into bootloader mode. Devices that support this
currently add a screenshot to the GsApp and then it gets shown before the update
is applied.

This isn't really suitable as when the update has been installed we then show
the screenshot prominently in the details page, which is really confusing.
It's a transient screenshot that's only requried for GsAppFlags _USER_ACTION so
give the property it deserves.