ci: Use malcontent subproject in the CI
Rather than downloading and building it separately. This makes no
functional difference to the CI, but simplifies the YML a little.

This requires adding the malcontent 0.8.0 dependencies to the Docker
images. The CI previously used malcontent 0.4.0, which is what
gnome-software depends on, but building it as a subproject requires some
fixes which are only available in 0.8.0.

Signed-off-by: Philip Withnall <withnall@endlessm.com>