ci: Cache subproject files in the CI images
This avoids them being re-downloaded on every CI run, which saves time
and bandwidth, and avoids the occasional spurious CI failure if they
can’t be downloaded because of network connectivity problems.

This requires bumping the Meson dependency on the Docker images to
0.50.0 for `meson subprojects download` and the fix in
https://github.com/mesonbuild/meson/commit/47b9c1a564756ac48a55da9a7c4d91787399c645.
The dependency of gnome-software outside the CI environment is unchanged.

Signed-off-by: Philip Withnall <withnall@endlessm.com>