ci: Add wrap file for libglib-testing
This is needed for building malcontent as a subproject; currently, Meson
doesn’t support recursively downloading the subprojects for nested wrap
files. (See https://github.com/mesonbuild/meson/issues/1489.) So we need
to provide `libglib-testing.wrap` in the top-level subprojects directory
in order for the `cache-subprojects.sh` script to correctly cache it in
the Docker images for CI.

This also requires bumping the malcontent dependency so that we pull in
a version of malcontent which uses the subproject.

Signed-off-by: Philip Withnall <withnall@endlessm.com>