flatpak: Update progress for some skipped ops
Currently we don't do any progress updates (gs_app_set_progress()) for
operations in a FlatpakTransaction that are skipped. This is problematic
in some situations such as when an app or runtime is up to date but its
locale extension needs an update to add a language (because the user
changed the system language since installing it). In these situations,
the app's progress is stuck at "Preparing..." forever and never gets to
1%. Fix the issue by updating progress for even skipped operations (in
which case the progress calculation excludes the size of the app itself
and only takes into account related operations' size).

Ideally we would test this situation in gs-self-test.c, which already
has a test for an extension update and connects to
notify::progress on the main app, but in practice adding
"g_assert_cmpint (progress_cnt, >, 0);" after "g_assert
(got_progress_installing);" leads to that assertion failing. It seems
the progress is only ever updated to GS_APP_PROGRESS_UNKNOWN even though
in actual use this works. I also tried adding 100 MB to extension update
in case the update was too small for progress to work, but still no
luck. gs_plugins_flatpak_app_update_func() has its assertion on
progress_cnt commented out, so I guess something is wrong even in the
case where there are no extensions.