flatpak: Set app to installed when a related thing is updated
When an app's locale extension is updated but the app is not, we put
progress updates on the app so they are user visible. However since
_transaction_operation_done() is only ever called for the locale, the
app never gets set to AS_APP_STATE_INSTALLED and instead gets frozen in
the "Preparing..." state when gs_plugin_loader_helper_free() sets the
progress to GS_APP_PROGRESS_UNKNOWN on it.

So set the app to installed when one of its related refs is updated, if
the app itself is being skipped.

Relatedly, update the unit test which tests extension updates so it
succeeds with Flatpak older than 1.7.3, because the changes in this
patch aren't active for those versions.