gs-details-page: Don’t list all OARS sections if app has ‘none’ rating
If an app has all its OARS metadata set to ‘none’ (i.e. no violence, no
inappropriate language, no unrestricted communications, etc.), show a
special message in the OARS popover, rather than an exhaustive list of
every OARS section.

This is only an issue with some content ratings schemes. Those, such as
PEGI, which have no age rating band which includes age 0 are not
affected. So this has to be tested with a system like INCAA or ESRB.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1008, #993