gs-details-page: Coalesce certain age ratings groups to clarify popover
There’s no point distinguishing between ‘app was rated as age X because
of sex-themes’ and ‘app was rated as age X because of sex-homosexuality’
when the two are identical in most jurisdictions. (Indeed,
discriminating by sexual orientation is illegal in many jurisdictions.)

Quite apart from legality, listing the two separately was confusing.

Instead, just mention `sex-themes` in the popover message, unless in a
jurisdiction where the rating for the two is different.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>