gs-details-page: Decide which content rating IDs to show based on age
Previously, the list of content rating IDs to highlight in the content
rating popover would be decided by whichever ones had the highest value
(intense, moderate, mild, none). However, it was possible for the CSM
age corresponding to a certain (ID, value) pair to be higher than the
age for a different ID with a higher value. This would result in a
relevant ID being missed off the list.

So convert the code to compare in terms of CSM ages, rather than the
content rating values which they are mapped from.

Continue checking for ‘none’ ratings using the `AsContentRatingValue`
rather than the CSM age, as it’s possible for the CSM age for
`AS_CONTENT_RATING_VALUE_NONE` to be >0.

An example of an app whose list of relevant IDs changes with this commit
is EDuke32, which goes from listing:
 * Rape or other violent sexual behavior
 * Prolonged nudity
 * Graphic sexual behavior
 * Strong or frequent use of profanity
 * Mature or sexual humour
 * Explicit discrimination based on gender, sexuality or religion
 * Graphic depictions of the act of prostitution
 * Overtly sexualised human characters
 * Depictions of bloodshed and the mutilation of body parts
 * Uncontrolled chat functionality between users
to listing only the ‘worst’ of those:
 * Rape or other violent sexual behaviour
 * Graphic depictions of the act of prostitution
 * Depictions of bloodshed and the mutilation of body parts

Spotted by Phaedrus Leeds.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>