flatpak: Fix displaying renamed apps
Currently for apps that have been renamed on Flathub, at least
org.gnome.iagno for example, the appstream data only exists under the
new name, so the app doesn't get properly displayed in the updates list
as it should so it can be updated to the new name. Instead there are
errors in the log "GsPluginFlatpak no match for ....: no results for
XPath query" and later "Gs app invalid as no name ..."

So use the appstream data under the new name in such circumstances. This
is not perfect in that we'll show the new name before the app has
actually been renamed, but it's certainly better than not showing the
app at all, and in practice renames are often just to change the app ID.

This makes renamed apps display properly in the Installed and Updates
lists, but there is still a remaining issue: after updating an app to
a new name, it doesn't show in the Installed list until after
gnome-software is closed and restarted.