fwupd: Use new API in fwupd 1.5.2 to download files
The fwupd project is soon to switch from libsoup to libcurl which means we can't
do the never-really-supported trick of stealing the SoupSession from the
FwupdClient anymore. Just use the new API provided to download files.

When we hard depend on this super-new version of fwupd we can drop a ton of
plugin code that now exists in libfwupd. We can also drop the internal export
of the SoupSession in the gnome-software plugins as the fwupd plugin was the
only consumer.