gs-details-page: Don’t show the origin box if not on the details page
The different repos an app can be installed from are loaded
asynchronously, but the widget for them (the origin box) is in a global
header bar. That means it’s possible for the asynchronous load to
complete after the user has switched away from the app details page, but
for the origin box to then still be shown.

Fix that by only showing the origin box if the details page is still
active when the asynchronous load completes.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #988