gs-category-page: Fix view scrolling if clicking tiles near the top
This was introduced in commit 997c6b5f and broken in/around commit
a8e93c9caa, and was supposed to scroll the
view to show the tile that had just been clicked on (or which was
focused by using the arrow keys on the keyboard). However, that
(specifically `gtk_container_set_focus_vadjustment()`) only works if the
child widgets are in the same coordinate system as the container. That’s
not the case here, as `featured_grid` (and other widgets) are between
`scrolledwindow_category` and `category_detail_box` in the widget
hierarchy. That means the scrolling jumps to the wrong place, and the
tile isn’t activated.

It might be possible to reimplement the mechanics of
`gtk_container_set_focus_vadjustment()` in gnome-software, but that
seems like a lot of fragile code for a small feature. Instead, just drop
it.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1040