gs-plugin-loader: Clear pending action on app once task is done
This typically manifests in the progress bar not disappearing from the
details page if installing an app fails.

Normally, when an app is installed, calling `gs_app_set_state()` to
change it to the `INSTALLED` state will clear the pending action.
However, if the installation fails early on, before the state change to
`INSTALLING`, the app’s state doesn’t need to be reverted, so its
pending action never gets cleared.

Since the pending action is originally set in
`gs_plugin_loader_schedule_task()`, it makes sense to clear that pending
action in `gs_plugin_loader_process_in_thread_pool_cb()`, which is the
paired function which actually runs the task.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #744