rpm-ostree: Don’t hang if the daemon crashes
If the rpm-ostreed daemon crashes during an operation, the rpm-ostree
plugin will continue to wait to receive a `Finished` signal from it,
forever.

Prevent that by watching the name owner of the proxy; if it becomes
unset, the daemon has disappeared for some reason, and the rpm-ostree
plugin should stop blocking.

Spotted while working on #1091, via code inspection. I haven’t attempted
to reproduce it in the wild.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>