gs-plugin-loader: Run each job thread with a new GMainContext
Otherwise the jobs will run with no thread-default main context, which
means that any main context iteration will typically iterate the global
default main context, and is likely to either cause code to execute on
the wrong thread (not thread safe), or to deadlock.

The general rule with GLib is that there should be exactly one,
distinct, `GMainContext` per thread.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1091