gs-application: Search for the installed application in app.launch
Clicking 'Launch' button in the "application was installed" bubble doesn't
launch just installed flatpak application. The problem is that the app.launch
is called only with the application ID, and then a GsApp instance is just
created from this ID, which is not enough to even refine the internal things,
because the app doesn't know what plugin it belongs to.

The fix is to:
 * pass both management plugin and the application ID to app.launch
 * search for the application by its ID
 * traverse returned list of found applications and find a matched installed application from the given plugin
 * launch the found application

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1103
Closes https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/555