flatpak: Improve performance of progress reporting
The `_ref_to_app()` function is called O(log N) times for each call to
`_transaction_progress_changed_cb()`, for each of the log N transaction
ops which are ancestors of the current op in the op tree. Each call to
`_ref_to_app()` results in a call to
`flatpak_installation_list_remote_refs_sync_full()`, which is not fast.

Cache the ref/app mappings so this doesn’t have to happen an order of
magnitude as often.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>