gs-feature-tile: Support RTL CSS for banners
If the feature tile is displayed in RTL mode, load any RTL CSS from the
`GnomeSoftware::FeatureTile-css-rtl` metadata key in preference to the
normal `GnomeSoftware::FeatureTile-css` key.

Reload the feature tile if the text direction changes.

This is a simpler approach than supporting `.class` and
`.class:dir(rtl)` CSS selectors in the same block of CSS. The problem
with allowing arbitrary blocks of CSS is that it’s very hard to validate
that they only do what they’re meant to, and don’t include CSS rules to
change the style of things they’re not meant to.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #773