gs-shell: Refactor mode change code to be event-based
Rather than making a load of changes to the shell widget state when
`gs_shell_change_mode()` is called, instead make the changes when the
shell’s stack changes page. `gs_shell_change_mode()` changes the page of
the stack, so this is equivalent to before.

This will allow (after further refactoring) for the shell’s stack to
have its page changed through other sources (by binding the
`visible-child-name` property to other controller widgets, for example)
and still have the UI work.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>