gs-shell: Use a final class instead of a derivable class
There’s no need for this to be derivable (it is likely never going to be
subclassed), so simplify the code and pointers a bit by making it a
final class instead.

The `priv->` → `shell->` replacements were done mechanically, and this
commit compiles, so shouldn’t need to be reviewed in detail.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>