gs-shell: Move various shell widget visibility changes out of pages
This code controls the visibility of shell-level widgets, so it should
be in the shell. It doesn’t really make sense for a page to control
whether it’s displayed with (for example) a menu button.

This is another step towards removing the layering violations which
allow the pages access to the shell’s UI.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>