build: Add a libgnomesoftware_dep internal dependency
This wraps up all the information about how to compile and link against
libgnomesoftware so that it can be used without separately listing the
includes, linker flags, built sources, etc.

This will be used in subsequent commits to tidy up the build system, and
add `glib-mkenums` support (which is not feasible without it).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>