lib: Use enum and flag types for GObject properties rather than guint64
This means that the varargs handling for `g_object_new()` should be able
to read the correct number of bytes off the stack to set the properties
correctly at construction time, regardless of whether built for i386 or
x86_64.

It will also improve type safety and the amount of information exposed
to language bindings.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1007