gs-details-page: Tweak presentation of add-ons
Tweak the display of add-ons to more closely match the design mock-ups
(see the ticket linked below). The mock-ups use a circular checkbox but
I'm not sure how to do that so it's still a square one.

Note, there is an assertion failure when the removal confirmation dialog
is created:
Gdk gdk_wayland_window_set_dbus_properties_libgtk_only: assertion 'GDK_IS_WAYLAND_WINDOW (window)' failed

...but this isn't unique to the removal of add-ons (it happens for apps
too), and it doesn't prevent the confirmation dialog or the removal from
happening.

Helps: #1111