lib: Build libgnomesoftware as a shared library as well as static
This means it can be linked to the plugins dynamically. This reduces the
installed plugin size by 50% (and correspondingly the disk reads and
memory for loading them). It doesn’t measurably affect startup time, and
overall reduces memory consumption by about 5% (although almost all of
that will be read-only mappings of the code).

Install the shared library in a private libdir to emphasise the fact
it’s not API-stable. This requires adding to the rpath for executables
which use it.

This changes the installation location for the static library to be in
the private libdir as well. The build processes of external plugins may
need to be modified to track this.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>