flatpak: Ensure ref details are set when refining renamed apps
The `FlatpakRef` metadata need to be set on a `GsApp` before it can be
refined to its renamed form. This happened when handling renames for
apps returned through normal code paths, but not for apps returned
through AppStream searches. In that case, the `GsApp` had come directly
from `gs_app_new()` inside the AppStream code, and not
`gs_flatpak_create_app()`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>