build: Move desktop category data into libgnomesoftware
This is one step towards removing the `desktop-categories` plugin and
hard-coding the desktop categories in gnome-software. It seems the
flexibility of defining the categories in a plugin has never been used,
so it will simplify the code (significantly) to define them statically
in the core. In particular, this means that they can be loaded
synchronously at startup, rather than asynchronously via
`gs_plugin_loader_job_get_categories_async()`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>